Dynamic poll interval
This commit modifies how the `Read` method polls records in the `command` table.

Previously, it simply hammered the database as frequently as possible to find new commands. This isn't
great as we're effectively choking the database for any other operations that may occur to do zero
work.

Now, I've added a dynamic backoff starting at a minimum value and capping out at a maximum. The
interval is doubled until it reaches the maximum. This allows the FSM to chill out a bit against
the database at times when there are no commands being written to the command table.

Signed-off-by: David Bond <davidsbond93@gmail.com>