Concurrency support
This commit provides a new option named SetConcurrency that allows the user of the package
to specify the number of queued commands to process concurrently. This effectively adds
batching with some slight differences.

Mainly, commands in a batch are all managed within a single database transaction, so are
dependent on each other's success. If any one command in the batch fails then all commands
in the batch will return to the queue. This makes idempotency quite important.

I wanted to add this kind of batching as I figured you could end up in situations where all
instances of the FSM get blocked waiting on a chain of dependent commands, so allowing nodes
to process large amounts of commands in a batch can help mitigate that problem.

Signed-off-by: David Bond <davidsbond93@gmail.com>